--- packages/dnb-eufemia/src/extensions/forms/Field/style/dnb-expiry.scss
@@ -1,52 +1,0 @@
-@import '../../../../style/core/utilities.scss';
-
-.dnb-forms-field-expiry {
-  // Hiding borders on sub inputs created by dnb-input__shell added by MaskedInput
-  .dnb-forms-field-expiry__input,
-  .dnb-forms-field-expiry__input:not([data-input-state='disabled']),
-  .dnb-forms-field-expiry__input:not([data-input-state='disabled']):hover,
-  .dnb-input.dnb-input__status--error.dnb-forms-field-expiry__input {
-    .dnb-forms-field-expiry__input--wrapper {
-      .dnb-input__shell {
-        --border-width: 0;
-        --border-color: transperant;
-        // Make background transperant to make outer dnb-input__shell border visible
-        background-color: transparent;
-      }
-    }
-  }
-
-  &__date-field {
-    &:first-of-type {
-      .dnb-forms-field-expiry__date-input {
-        padding-right: 0.125rem;
-        padding-left: 0.8rem;
-      }
-    }
-
-    &:last-of-type {
-      .dnb-forms-field-expiry__date-input {
-        padding-left: 0.125rem;
-        padding-right: 0.8rem;
-      }
-    }
-  }
-
-  &:not(#{&}__status--error) &__date-field {
-    color: var(--color-black-55);
-  }
-
-  &:not(#{&}__status--error) & {
-    &__date-field--has-value {
-      &:not([data-input-state='disabled']) {
-        color: var(--color-black);
-      }
-    }
-  }
-
-  .dnb-forms-field-expiry__seperator {
-    &--no-highlight {
-      color: var(--color-black-55);
-    }
-  }
-}
